refactor(api): move usecase
get-current-user
